updated props with example
