+++ bin/example_props/basic_sample_props.json
