+++ bin/example_props/detail_sample_props.json
